# docs/swagger/stassosrv-1.0.0-swagger.yaml
swagger: '2.0'
info:
  description: API componente staaudit
  version: 1.0.0
  title: STASSOSRV
  contact:
    email: andrea.fontanelli@eng.it
# tags are used for organizing operations
tags:
- name: test
  description: test resources
- name: sso
  description: Servizio per la gestione dell'autorizzazione e profilazione utente


# basePath: /api
# schemes:
#    - http

paths:
  /test/resources:
    get:
      tags:
        - test
      summary: Test resources
      description: Effettua il test dei servizi richiamati
      operationId: testResources
      produces:
      - application/json
      responses:
        200:
          schema:
            $ref: '#/definitions/GenericResponse'
          description: OK
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorDto'
  /sso/utente:
    get:
      tags:
        - sso
      summary: recupera l'utente
      description: recupera l'utente dall'identità iride e il codice applicazione
      operationId: getUtente
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: query
          name: identita
          type: string
          required: true
          description: Identità restituita dal SSO SHIBB
      responses:
        200:
          schema:
            $ref: '#/definitions/GenericResponse'
          description: OK
        500:
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorDto'
definitions:
  Utente:
    type: object
    description: Utente
    properties:
      id:
        type: string
        description: Id Utente
        example: 'Identificativo Utente'
      tipoUtente:
        $ref: '#/definitions/TipoUtente'
      ente:
        $ref: '#/definitions/Ente'
      elencoCassa:
        type: array
        description: Elenco casse
        items:
          $ref: '#/definitions/Cassa'
      elencoSportello:
        type: array
        description: Elenco sportelli
        items:
          $ref: '#/definitions/Sportello'
  TipoUtente:
    type: object
    description: Tipo Utente
    properties:
      codice:
        type: string
        description: Codice Utente
        example: 'OP-Operatore, AP-Applicativo, BA-Batch'
      descrizione:
        type: string
        description: Descrizione tipo Utente
        example: 'OP-Operatore, AP-Applicativo, BA-Batch'
  Ente:
    type: object
    description: Ente
    properties:
      email:
        type: string
        description: Email
        example: 'Email Ente'
      indirizzo:
        type: string
        description: Indirizzo
        example: 'Indirizzo Ente'
      civico:
        type: string
        description: Civico
        example: 'Civico Ente'
      cap:
        type: string
        description: CAP
      comune:
        type: string
        description: Comune
      codiceFiscale:
        type: string
        description: Comune
      provincia:
        type: string
        description: Provincia
  Abilitazione:
    type: object
    description: Abilitazione
    properties:
      codice:
        type: string
        description: Codice Utente
        example: 'OP-Operatore, AP-Applicativo, BA-Batch'
  Cassa:
    type: object
    description: Cassa
    properties:
      codice:
        type: string
        description: Codice Utente
        example: 'OP-Operatore, AP-Applicativo, BA-Batch'
  Sportello:
    type: object
    description: Cassa
    properties:
      codice:
        type: string
        description: Codice Utente
        example: 'OP-Operatore, AP-Applicativo, BA-Batch'
  GenericResponse:
    type: object
    description: Risposta di un servizio generico
    properties:
      message:
        type: string
        description: Contiene il messaggio eventuale del servizio chiamato
        example: 'Servizio ha risposto correttamente'
  ErrorDto:
    type: object
    description: Rappresenta l'errore restituito da servizio
    properties:
      status:
        type: integer
        description: status http restituito
        example: '500, 400, ecc'
      code:
        type: string
        description: code http restituito
        example: 'Internal server error, ecc'
      description:
        type: string
        description: la descrizione dell'errore
        example: 'Internal server error, ecc'
      errorDetails:
        type: array
        description: i dettagli errore
        items:
          $ref: '#/definitions/ErrorDetail'
  ErrorDetail:
    type: object
    description: rappresenta il dettaglio errore restituito in caso di errori di validazione sui campi di input
    properties:
      field:
        type: string
        description: il nome del campo per cui è falita la validazione
        example: 'name, surname'
      error:
        type: string
        description: l'errore di validazione effettivo
        example: 'campo obbligatorio, ecc'
      description:
        type: string
        description: la descrizione dell'errore
        example: 'Internal server error, ecc'
  
# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
# basePath: /studiofontanelli/staaudit/1.0.0
schemes:
 - https
 - http
# Added by API Auto Mocking Plugin
basePath: /andreafontanelli/stassosrv/1.0.0